moving functions to more appropriate modules
