+++ cionic/api.py
@@ -11,10 +11,9 @@ import sys
 import zipfile
 
 import numpy as np
-import pandas as pd
 import requests
 
-from cionic import json2npy, segmenter, tools
+from cionic import json2npy, npz_utils, segmenter, tools
 
 apiver = '0.22'
 server = None
@@ -389,10 +388,6 @@ def download_files(urlpath, directory, include=None, exclude=None, ver=apiver):
     return results
 
 
-# TODO: fix signal.ipynb
-# TODO: fix gait.ipynb
-
-
 def add_arrays_to_npz_and_store(
     npz: np.lib.npyio.NpzFile,
     array_dict: dict[str, np.ndarray],
@@ -443,132 +438,6 @@ def add_arrays_to_npz_and_store(
                     np.save(fp, arr, allow_pickle=False)
 
 
-def get_limb_eulers(
-    npz: np.lib.npyio.NpzFile, degrees: bool = True
-) -> tuple[dict[str, np.ndarray], np.ndarray]:
-    """
-    Extract Euler angles from quaternion segments in a .npz file.
-
-    Args:
-        npz (np.lib.npyio.NpzFile): Opened .npz file containing segment data.
-        degrees (bool, optional): If True, returns Euler angles in degrees.
-
-    Returns:
-        tuple:
-            - dict[str, np.ndarray]: Dict of euler_path names to Euler arrays.
-            - np.ndarray: Array of new segment metadata dicts for the Euler streams.
-    """
-    print("getting limb eulers from npz", file=sys.stderr)
-    limb_eulers = {}
-    new_limb_segments = []
-    for seg in change_segments_column_dtype(npz['segments']):
-        if seg['stream'] != 'fquat':
-            continue
-        stream = tools.stream_quat2euler(
-            stream=npz[seg['path']], calibration=seg['calibration'], degrees=degrees
-        )
-        euler_path = f'{seg["path"]}2euler'
-        limb_eulers[euler_path] = stream
-
-        new_segment = seg.copy()
-        new_segment['path'] = euler_path
-        new_segment['fields'] = 'x y z'
-        new_segment['stream'] = 'euler'
-        new_limb_segments.append(new_segment)
-
-    return limb_eulers, np.array(new_limb_segments)
-
-
-def pandas_to_ndarray(df: pd.DataFrame) -> np.ndarray:
-    """
-    Convert a pandas DataFrame to a NumPy ndarray.
-
-    Args:
-        df (pandas.DataFrame): The DataFrame to convert.
-
-    Returns:
-        np.ndarray: The converted ndarray.
-    """
-    array = np.array(
-        list(df.itertuples(index=False)),
-        dtype=np.dtype(
-            {
-                'names': df.columns.tolist(),
-                'formats': [df[col].dtype for col in df.columns],
-            }
-        ),
-    )
-    return array
-
-
-def get_joint_eulers(
-    npz: np.lib.npyio.NpzFile,
-) -> tuple[dict[str, np.ndarray], list[np.ndarray]]:
-    '''
-    Extracts joint Euler angle data and corresponding segment information from NPZ.
-
-    Args:
-        npz (np.lib.npyio.NpzFile): The NPZ file containing joint data.
-
-    Returns:
-        tuple:
-            - joint_eulers (dict): A dictionary mapping each joint stream path to its
-              Euler angle data as a NumPy ndarray.
-            - new_joint_segments (list): A list of segment metadata arrays.
-    '''
-    print("getting joint eulers from npz", file=sys.stderr)
-    segments = change_segments_column_dtype(npz['segments'])
-    streams_data_packet = tools.get_joint_streams(npz, segmented=False)
-    joint_eulers = {}
-    new_joint_segments = []
-
-    for stream_data in streams_data_packet:
-        data_stream = stream_data['data_stream']
-        joint_eulers[stream_data['path']] = pandas_to_ndarray(data_stream)
-
-        seg_dtype = segments.dtype
-        values = tuple(stream_data.get(name, '') for name in seg_dtype.names)
-
-        new_segment = np.array([values], dtype=seg_dtype)[0]
-        new_joint_segments.append(new_segment)
-
-    return joint_eulers, new_joint_segments
-
-
-def change_segments_column_dtype(segments: np.ndarray, dtype_dict=None) -> np.ndarray:
-    '''
-    Change dtype of specified columns in a structured numpy array.
-
-    Args:
-        segments (np.ndarray): Structured numpy array (like pandas DataFrame).
-        new_dtypes (list of tuples): List of (field, dtype) to update.
-
-    Returns:
-        np.ndarray: New array with updated dtypes.
-    '''
-    if dtype_dict is None:
-        dtype_dict = {
-            'position': 'U20',
-            'device': 'U40',
-            'path': 'U100',
-        }
-
-    # Build new dtype: update only specified fields, keep others the same
-    new_dtype = []
-    for name, oldtype in segments.dtype.descr:
-        if name in dtype_dict.keys():
-            new_dtype.append((name, dtype_dict[name]))
-        else:
-            new_dtype.append((name, oldtype))
-
-    # Create new array and copy data
-    new_segments = np.empty(segments.shape, dtype=new_dtype)
-    for name in segments.dtype.names:
-        new_segments[name] = segments[name]
-
-    return new_segments
-
-
 def include_eulers_to_npz(destpath: str) -> None:
     """
     Load a .npz file, compute limb and joint Euler angles, update segments,
@@ -583,12 +452,12 @@ def include_eulers_to_npz(destpath: str) -> None:
         print(f"File {destpath} not found.", file=sys.stderr)
         return
 
-    limb_eulers, new_limb_segments = get_limb_eulers(npz)
-    joint_eulers, new_joint_segments = get_joint_eulers(npz)
+    limb_eulers, new_limb_segments = tools.get_limb_eulers(npz)
+    joint_eulers, new_joint_segments = tools.get_joint_eulers(npz)
 
     updated_segments = np.concatenate(
         [
-            change_segments_column_dtype(npz['segments']),
+            npz_utils.change_segments_column_dtype(npz['segments']),
             new_limb_segments,
             new_joint_segments,
         ]
